+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 ### Added
+
 - Support for rackup files
+- Settings for 'quote_style' to settings.md
 
 ## [0.7.0] - 2019-04-28