+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Fixed
+
+### Changed
+
+### Added
+
+## [0.8.0] - 2020-01-02
+
+### Fixed
 - Handle endless ranges in ruby 2.6
 - Handle empty string hash keys